test(user): move UserProfileContainer behavior tests, slim UserGetContainer test to id passthrough
